Omogucen pregled sadrzaja odgovora na selektovanu zalbu

+++ backend/src/main/java/mrsisa/projekat/zalba/ZalbaRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ZalbaRepository extends JpaRepository<Zalba, Long> {
+public interface ZalbaRepository extends JpaRepository<Zalba, Integer> {
     List<Zalba> findAllByPacijent(String pacijent);
+
 }
